+++ apps/repository/js/splash/splashes/whats_new/WhatsNewRef.tsx
@@ -9,16 +9,23 @@ import {LifecycleToggle} from '../../../../../../web/js/ui/util/LifecycleToggle'
 import {LifecycleEvents} from '../../../../../../web/js/ui/util/LifecycleEvents';
 import * as semver from 'semver';
 import {DatastoreOverview} from '../../../../../../web/js/datastore/Datastore';
+import {DatastoreOverviewPolicies} from '../DatastoreOverviewPolicies';
 
 export class WhatsNewRef implements PrioritizedComponentRef {
 
     public readonly id = 'whats-new';
 
     constructor() {
-        this.priority = Providers.memoize(this.priority.bind(this));
+        // this.priority = Providers.memoize(this.priority.bind(this));
+        // this.priority = Providers.memoize(this.priority.bind(this));
     }
 
     public priority(datastoreOverview: DatastoreOverview): number | undefined {
+
+        if (! DatastoreOverviewPolicies.isLevel('24h', datastoreOverview)) {
+            // must be running for at least 24 hours.
+            return undefined;
+        }
 
         const hasTourTerminated = this.hasTourTerminated();